Add Book Feedback BookTransactionHistory entities and the relationships between them
