--- book-network/src/main/java/com/ka/book/BookNetworkApiApplication.java
@@ -1,31 +1,0 @@
-package com.ka.book;
-
-import com.ka.book.role.Role;
-import com.ka.book.role.RoleRepository;
-import org.springframework.boot.CommandLineRunner;
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.context.annotation.Bean;
-import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
-import org.springframework.scheduling.annotation.EnableAsync;
-
-@SpringBootApplication
-@EnableJpaAuditing
-@EnableAsync
-public class BookNetworkApiApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(BookNetworkApiApplication.class, args);
-    }
-
-    @Bean
-    public CommandLineRunner runner(RoleRepository roleRepository) {
-        return args -> {
-            if (roleRepository.findByName("USER").isEmpty()) {
-                roleRepository.save(
-                        Role.builder().name("USER").build()
-                );
-            }
-        };
-    }
-}
