--- book-network/src/main/java/com/ka/book/auth/AuthenticationResponse.java
@@ -1,12 +1,0 @@
-package com.ka.book.auth;
-
-import lombok.Builder;
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-@Builder
-public class AuthenticationResponse {
-    private String token;
-}
