--- book-network/src/main/java/com/ka/book/auth/AuthenticationService.java
@@ -1,128 +1,0 @@
-package com.ka.book.auth;
-
-
-import com.ka.book.email.EmailDetails;
-import com.ka.book.email.EmailService;
-import com.ka.book.email.EmailTemplateName;
-import com.ka.book.role.RoleRepository;
-import com.ka.book.security.JwtService;
-import com.ka.book.user.Token;
-import com.ka.book.user.TokenRepository;
-import com.ka.book.user.User;
-import com.ka.book.user.UserRepository;
-import jakarta.mail.MessagingException;
-import lombok.RequiredArgsConstructor;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.security.authentication.AuthenticationManager;
-import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
-import org.springframework.security.core.userdetails.UsernameNotFoundException;
-import org.springframework.security.crypto.password.PasswordEncoder;
-import org.springframework.stereotype.Service;
-
-import java.security.SecureRandom;
-import java.time.LocalDateTime;
-import java.util.HashMap;
-import java.util.List;
-
-@Service
-@RequiredArgsConstructor
-public class AuthenticationService {
-
-    private final RoleRepository roleRepository;
-    private final PasswordEncoder passwordEncoder;
-    private final UserRepository userRepository;
-    private final TokenRepository tokenRepository;
-    private final EmailService emailService;
-    private final AuthenticationManager authenticationManager;
-    private final JwtService jwtService;
-    @Value("${application.mailing.frontend.activation-url}")
-    private String activationUrl;
-
-    public void register(RegistrationRequest request) throws MessagingException {
-        var userRole = roleRepository.findByName("USER")
-                //TODO Exception handling
-                .orElseThrow(() -> new IllegalStateException("Role USER is not initialized"));
-        var user = User.builder()
-                .firstName(request.getFirstName())
-                .lastName(request.getLastName())
-                .email(request.getEmail())
-                .password(passwordEncoder.encode(request.getPassword()))
-                .accountLocked(false)
-                .enabled(false)
-                .roles(List.of(userRole))
-                .build();
-        userRepository.save(user);
-        sendValidationEmail(user);
-    }
-
-    private void sendValidationEmail(User user) throws MessagingException {
-        var newToken = generateAndSaveActivationToken(user);
-        EmailDetails emailDetails = new EmailDetails(
-                user.getEmail(),
-                user.fullName(),
-                EmailTemplateName.ACTIVATE_ACCOUNT,
-                activationUrl,
-                newToken,
-                "Account Activation"
-        );
-        emailService.sendEmail(emailDetails);
-    }
-
-    private String generateAndSaveActivationToken(User user) {
-
-        // generate token
-        String generatedToken = generateActivationCode(6);
-        var token = Token.builder()
-                .token(generatedToken)
-                .createdAt(LocalDateTime.now())
-                .expiresAt(LocalDateTime.now().plusMinutes(15))
-                .user(user)
-                .build();
-        tokenRepository.save(token);
-        return generatedToken;
-    }
-
-    private String generateActivationCode(int length) {
-        String characters = "0123456789";
-        StringBuilder codeBuilder = new StringBuilder();
-        SecureRandom secureRandom = new SecureRandom();
-        for(int i = 0; i < length; i++){
-            int randomIndex = secureRandom.nextInt(characters.length());
-            codeBuilder.append(characters.charAt(randomIndex));
-        }
-        return codeBuilder.toString();
-    }
-
-    public AuthenticationResponse authenticate(AuthenticationRequest request) {
-        var auth = authenticationManager.authenticate(
-                new UsernamePasswordAuthenticationToken(
-                        request.getEmail(),
-                        request.getPassword()
-                )
-        );
-        var claims = new HashMap<String, Object>();
-        var user = ((User) auth.getPrincipal());
-        claims.put("fullname", user.fullName());
-        var jwtToken = jwtService.generateToken(claims, user);
-        return AuthenticationResponse.builder()
-                .token(jwtToken).build();
-    }
-
-    public void activateAccount(String token) throws MessagingException {
-        Token savedToken = tokenRepository.findByToken(token)
-                // TODO Handle Exception
-                .orElseThrow(() -> new RuntimeException("Invalid token"));
-        if(LocalDateTime.now().isAfter(savedToken.getExpiresAt()))
-        {
-            sendValidationEmail(savedToken.getUser());
-            throw new RuntimeException("Activation token has expired. " +
-                    "A new token has been sent to the same email address!");
-        }
-        var user = userRepository.findById(savedToken.getUser().getId())
-                .orElseThrow(() -> new UsernameNotFoundException("User not found"));
-        user.setEnabled(true);
-        userRepository.save(user);
-        savedToken.setValidatedAt(LocalDateTime.now());
-        tokenRepository.save(savedToken);
-    }
-}
