--- book-network/src/main/java/com/ka/book/email/EmailDetails.java
@@ -1,10 +1,0 @@
-package com.ka.book.email;
-
-public record EmailDetails(
-        String to,
-        String username,
-        EmailTemplateName emailTemplate,
-        String confirmationUrl,
-        String activationCode,
-        String subject
-) {}
