--- book-network/src/main/java/com/ka/book/email/EmailTemplateName.java
@@ -1,15 +1,0 @@
-package com.ka.book.email;
-
-import lombok.Getter;
-
-@Getter
-public enum EmailTemplateName {
-
-    ACTIVATE_ACCOUNT("activate_account");
-
-    private final String name;
-
-    EmailTemplateName(String name) {
-        this.name = name;
-    }
-}
