--- book-network/src/main/java/com/ka/book/handler/BusinessErrorCodes.java
@@ -1,32 +1,0 @@
-package com.ka.book.handler;
-
-import lombok.Getter;
-import org.springframework.http.HttpStatus;
-
-import static org.springframework.http.HttpStatus.FORBIDDEN;
-import static org.springframework.http.HttpStatus.NOT_IMPLEMENTED;
-
-
-public enum BusinessErrorCodes {
-
-    NO_CODE(0, NOT_IMPLEMENTED, "No code"),
-    INCORRECT_CURRENT_PASSWORD(300, HttpStatus.BAD_REQUEST, "Current password is incorrect"),
-    NEW_PASSWORD_DOES_NOT_MATCH(301, HttpStatus.BAD_REQUEST, "The new password does not match"),
-    ACCOUNT_LOCKED(302, FORBIDDEN, "User account is locked"),
-    ACCOUNT_DISABLED(303, FORBIDDEN, "User account is disabled"),
-    BAD_CREDENTIALS(304, FORBIDDEN, "Login and/or password is incorrect")
-    ;
-
-    @Getter
-    private final int code;
-    @Getter
-    private final String description;
-    @Getter
-    private final HttpStatus httpStatus;
-
-    BusinessErrorCodes(int code, HttpStatus httpStatus, String description) {
-        this.code = code;
-        this.description = description;
-        this.httpStatus = httpStatus;
-    }
-}
