--- book-network/src/main/java/com/ka/book/handler/ExceptionResponse.java
@@ -1,23 +1,0 @@
-package com.ka.book.handler;
-
-
-import com.fasterxml.jackson.annotation.JsonInclude;
-import lombok.*;
-
-import java.util.Map;
-import java.util.Set;
-
-@Getter
-@Setter
-@Builder
-@AllArgsConstructor
-@NoArgsConstructor
-@JsonInclude(JsonInclude.Include.NON_EMPTY)
-public class ExceptionResponse {
-
-    private Integer businessErrorCode;
-    private String businessErrorDescription;
-    private String error;
-    private Set<String> validationErrors;
-    private Map<String, String> errors;
-}
