--- book-network/src/main/java/com/ka/book/role/RoleRepository.java
@@ -1,9 +1,0 @@
-package com.ka.book.role;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.Optional;
-
-public interface RoleRepository extends JpaRepository<Role, Long> {
-    Optional<Role> findByName(String role);
-}
