--- book-network/src/main/java/com/ka/book/user/Token.java
@@ -1,26 +1,0 @@
-package com.ka.book.user;
-
-import jakarta.persistence.*;
-import lombok.*;
-
-import java.time.LocalDateTime;
-
-@Getter
-@Setter
-@Builder
-@AllArgsConstructor
-@NoArgsConstructor
-@Entity
-public class Token {
-    @Id
-    @GeneratedValue
-    private Long id;
-    private String token;
-    private LocalDateTime createdAt;
-    private LocalDateTime expiresAt;
-    private LocalDateTime validatedAt;
-
-    @ManyToOne
-    @JoinColumn(name = "userId", nullable = false)
-    private User user;
-}
