--- book-network/src/main/java/com/ka/book/user/TokenRepository.java
@@ -1,9 +1,0 @@
-package com.ka.book.user;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.Optional;
-
-public interface TokenRepository extends JpaRepository<Token, Long> {
-    Optional<Token> findByToken(String token);
-}
