--- book-network/src/main/java/com/ka/book/user/UserRepository.java
@@ -1,9 +1,0 @@
-package com.ka.book.user;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.Optional;
-
-public interface UserRepository extends JpaRepository<User, Long> {
-    Optional<User> findByEmail(String email);
-}
