+++ book-network/src/main/java/com/ka/bsn/auth/AuthenticationResponse.java
@@ -1,0 +1,12 @@
+package com.ka.bsn.auth;
+
+import lombok.Builder;
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+@Builder
+public class AuthenticationResponse {
+    private String token;
+}
