+++ book-network/src/main/java/com/ka/bsn/auth/RegistrationRequest.java
@@ -1,0 +1,29 @@
+package com.ka.bsn.auth;
+
+import jakarta.validation.constraints.Email;
+import jakarta.validation.constraints.NotBlank;
+import jakarta.validation.constraints.NotEmpty;
+import jakarta.validation.constraints.Size;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+@Builder
+public class RegistrationRequest {
+    @NotEmpty(message = "First name is mandatory")
+    @NotBlank(message = "First name is mandatory")
+    private String firstName;
+    @NotEmpty(message = "Last name is mandatory")
+    @NotBlank(message = "Last name is mandatory")
+    private String lastName;
+    @Email(message = "Invalid email format")
+    @NotEmpty(message = "Email is mandatory")
+    @NotBlank(message = "Email is mandatory")
+    private String email;
+    @NotEmpty(message = "Password is mandatory")
+    @NotBlank(message = "Password is mandatory")
+    @Size(min = 8, message = "Password should have at least 8 characters")
+    private String password;
+}
