+++ book-network/src/main/java/com/ka/bsn/book/Book.java
@@ -1,0 +1,38 @@
+package com.ka.bsn.book;
+
+import com.ka.bsn.common.BaseEntity;
+import com.ka.bsn.feedback.Feedback;
+import com.ka.bsn.history.BookTransactionHistory;
+import com.ka.bsn.user.User;
+import jakarta.persistence.*;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+import lombok.experimental.SuperBuilder;
+
+import java.util.List;
+
+@Entity
+@Getter
+@Setter
+@SuperBuilder
+@AllArgsConstructor
+@NoArgsConstructor
+public class Book extends BaseEntity {
+    private String title;
+    private String authorName;
+    private String isbn;
+    private String synopsis;
+    private String bookCover;
+    private boolean archived;
+    private boolean sharable;
+
+    @ManyToOne
+    @JoinColumn(name = "owner_id")
+    private User owner;
+    @OneToMany(mappedBy = "book")
+    private List<Feedback> feedbacks;
+    @OneToMany(mappedBy = "book")
+    private List<BookTransactionHistory> histories;
+}
