+++ book-network/src/main/java/com/ka/bsn/email/EmailDetails.java
@@ -1,0 +1,10 @@
+package com.ka.bsn.email;
+
+public record EmailDetails(
+        String to,
+        String username,
+        EmailTemplateName emailTemplate,
+        String confirmationUrl,
+        String activationCode,
+        String subject
+) {}
