+++ book-network/src/main/java/com/ka/bsn/email/EmailTemplateName.java
@@ -1,0 +1,15 @@
+package com.ka.bsn.email;
+
+import lombok.Getter;
+
+@Getter
+public enum EmailTemplateName {
+
+    ACTIVATE_ACCOUNT("activate_account");
+
+    private final String name;
+
+    EmailTemplateName(String name) {
+        this.name = name;
+    }
+}
