+++ book-network/src/main/java/com/ka/bsn/handler/ExceptionResponse.java
@@ -1,0 +1,23 @@
+package com.ka.bsn.handler;
+
+
+import com.fasterxml.jackson.annotation.JsonInclude;
+import lombok.*;
+
+import java.util.Map;
+import java.util.Set;
+
+@Getter
+@Setter
+@Builder
+@AllArgsConstructor
+@NoArgsConstructor
+@JsonInclude(JsonInclude.Include.NON_EMPTY)
+public class ExceptionResponse {
+
+    private Integer businessErrorCode;
+    private String businessErrorDescription;
+    private String error;
+    private Set<String> validationErrors;
+    private Map<String, String> errors;
+}
