+++ book-network/src/main/java/com/ka/bsn/history/BookTransactionHistory.java
@@ -1,0 +1,33 @@
+package com.ka.bsn.history;
+
+import com.ka.bsn.book.Book;
+import com.ka.bsn.common.BaseEntity;
+import com.ka.bsn.user.User;
+import jakarta.persistence.Entity;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+import lombok.experimental.SuperBuilder;
+
+@Entity
+@Getter
+@Setter
+@SuperBuilder
+@AllArgsConstructor
+@NoArgsConstructor
+public class BookTransactionHistory extends BaseEntity {
+
+    @ManyToOne
+    @JoinColumn(name = "user_id")
+    private User user;
+    @ManyToOne
+    @JoinColumn(name = "book_id")
+    private Book book;
+
+    private boolean returned;
+    private boolean returnApproved;
+
+}
