+++ book-network/src/main/java/com/ka/bsn/role/Role.java
@@ -1,0 +1,40 @@
+package com.ka.bsn.role;
+
+
+import com.fasterxml.jackson.annotation.JsonIgnore;
+import com.ka.bsn.user.User;
+import jakarta.persistence.*;
+import lombok.*;
+import org.springframework.data.annotation.CreatedDate;
+import org.springframework.data.annotation.LastModifiedDate;
+import org.springframework.data.jpa.domain.support.AuditingEntityListener;
+
+import java.time.LocalDateTime;
+import java.util.List;
+
+@Getter
+@Setter
+@Builder
+@AllArgsConstructor
+@NoArgsConstructor
+@Entity
+@EntityListeners(AuditingEntityListener.class)
+public class Role {
+
+    @Id
+    @GeneratedValue
+    private Long id;
+    @Column(unique = true)
+    private String name;
+
+    @ManyToMany(mappedBy = "roles")
+    @JsonIgnore
+    private List<User> users;
+
+    @CreatedDate
+    @Column(nullable = false, updatable = false)
+    private LocalDateTime createdDate;
+    @LastModifiedDate
+    @Column(insertable = false)
+    private LocalDateTime lastModifiedDate;
+}
