+++ book-network/src/main/java/com/ka/bsn/role/RoleRepository.java
@@ -1,0 +1,9 @@
+package com.ka.bsn.role;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.Optional;
+
+public interface RoleRepository extends JpaRepository<Role, Long> {
+    Optional<Role> findByName(String role);
+}
