+++ book-network/src/main/java/com/ka/bsn/user/TokenRepository.java
@@ -1,0 +1,9 @@
+package com.ka.bsn.user;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.Optional;
+
+public interface TokenRepository extends JpaRepository<Token, Long> {
+    Optional<Token> findByToken(String token);
+}
