+++ book-network/src/main/java/com/ka/bsn/user/UserRepository.java
@@ -1,0 +1,9 @@
+package com.ka.bsn.user;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.Optional;
+
+public interface UserRepository extends JpaRepository<User, Long> {
+    Optional<User> findByEmail(String email);
+}
