--- book-network/src/test/java/com/ka/book/BookNetworkApiApplicationTests.java
@@ -1,13 +1,0 @@
-package com.ka.book;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class BookNetworkApiApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
