+++ book-network/src/test/java/com/ka/bsn/BookNetworkApiApplicationTests.java
@@ -1,0 +1,13 @@
+package com.ka.bsn;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class BookNetworkApiApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
